kexec/arm64: Add support for handling zlib compressed (Image.gz) image Currently the kexec_file_load() support for arm64 doesn't allow handling zlib compressed (i.e. Image.gz) image.
Since most distributions use 'make zinstall' rule inside
'arch/arm64/boot/Makefile' to install the arm64
Image.gz compressed file inside the boot destination directory (for e.g.
/boot), currently we cannot use kexec_file_load() to load vmlinuz (or
Image.gz):

 # file /boot/vmlinuz
   /boot/vmlinuz: gzip compressed data, was "Image", <..snip..>, max
   compression, from Unix, original size 21945120

 Now, since via kexec_file_load() we pass the 'fd' of Image.gz
 (compressed file) via the following command line ...

 # kexec -s -l /boot/vmlinuz-`uname -r` --initrd=/boot/initramfs-`uname
   -r`.img --reuse-cmdline

... kernel returns -EINVAL error value, as it is not able to locate
the magic number =0x644d5241, which is expected in the 64-byte header
of the decompressed kernel image.

We can fix this in user-space kexec-tools, which handles an
'Image.gz' being passed via kexec_file_load(), using an approach
as follows:

a). Copy the contents of Image.gz to a temporary file.
b). Decompress (gunzip-decompress) the contents inside the
    temporary file.
c). Pass the 'fd' of the temporary file to the kernel space. So
    basically the kernel space still gets a decompressed kernel
    image to load via kexec-tools

I tested this patch for the following three use-cases:

1. Uncompressed Image file:
 #kexec -s -l Image --initrd=/boot/initramfs-`uname -r`.img --reuse-cmdline

2. Signed Image file:
 #kexec -s -l Image.signed --initrd=/boot/initramfs-`uname -r`.img --reuse-cmdline

3. zlib compressed Image.gz file:
 #kexec -s -l /boot/vmlinuz-`uname -r` --initrd=/boot/initramfs-`uname -r`.img --reuse-cmdline

kexec/kexec-zlib.h: Add 'is_zlib_file()' helper function

   This patch adds 'is_zlib_file()' helper function which can be
   used to quickly determine with the passed kernel image is a zlib
   compressed kernel image.

This is specifically useful for arm64 zImage (or Image.gz) support,
which is introduced by later patches in this patchset.

[PATCH] arm64: Disable PIC

  Fix
   | cc1: sorry, unimplemented: code model 'large' with -fPIC
   | make: *** [Makefile:118: purgatory/arch/arm64/entry.o] Error 1

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Bhupesh Sharma <bhsharma@redhat.com>
Signed-off-by: Simon Horman <horms@verge.net.au>